correlation with prop. and clr
